theBot/requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
platformdirs==4.3.6
plotly==6.0.1
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
setuptools==76.0.0
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
boto3==1.38.1
botocore==1.38.1
watchtower==3.4.0
colorlog==6.9.0